refactor(GCS): Migrate to using mockery/testify
This is my first attempt at using the [mockery](https://vektra.github.io/mockery/latest/) library to generate mocks.
Previously we crafted our own mocks by hand for the GCS
StorageClientInterface and hand wrote our own interceptors. I generated
the mocked interfaces by running `mockery` locally.
In order for `mockery` to recognize files, I had to create a `main.go` file in the root of the project.
This newer way of running `mockery` requires you to generate mocks for
_all_ interfaces.

I updated the tests for buckets to use the new mocked interface.
On Monday my plan is to extend the mocks and use testify to alter the
behavior so I can do things like handle errors, and all of that fun
stuff.