feat(gcs): Implement tests for collector
This introduces tests that are aimed to validate that GCS collector emits the proper tests after Collect is called.
Right now this just has a single case which is run all of the things.
I'd like to figure out a way to setup multiple cases, but a more pressing problem exists.
If you run all of the tests(`go test ./...`) everything passes.
However if you just run `TestCollector_Collect` it will fail with the following error:

```
Diff:
        	            	--- metric output does not match expectation; want
        	            	+++ got:
        	            	@@ -4,5 +4,2 @@
        	            	 gcp_gcs_bucket_info{bucket_name="testbucket-2",location="us",location_type="multi-region",storage_class="STANDARD"} 1
        	            	-# HELP gcp_gcs_operations_cost GCS operations cost per 1k requests
        	            	-# TYPE gcp_gcs_operations_cost gauge
        	            	-gcp_gcs_operations_cost{location="us-east1",opclass="",storage_class=""} 0.004
        	            	 # HELP gcp_gcs_operations_discount GCS operations discount
        	            	@@ -68,5 +65,2 @@
        	            	 gcp_gcs_storage_discount{location="us-east1",storage_class="STANDARD"} 0
        	            	-# HELP gcp_gcs_storage_hourly_cost GCS storage hourly cost in GiB
        	            	-# TYPE gcp_gcs_storage_hourly_cost gauge
        	            	-gcp_gcs_storage_hourly_cost{location="us-east1",storage_class=""} 5.376344086021506e-06
```

My hunch is that similar to AWS, we have a scenario where the global variables are instantiated between runs and persist there state, leading a flaky test.

I'll introduce one more commit to refactor the metrics into a Metrics struct so we can start off each test with a fresh suite of metrics.