feat(gcs): Create unit tests + benchmarks (#42)
Introduces unit tests and benchmarks for the bucket_cache.go file.
It was raised during implementation that using defer has a  performance penalty, but I wanted to use benchmarks to measure it. 
By removing defers, we end up with roughly 5% performance gains.
In practice though, this should be minor and have negligible impact on overall performance.