feat(google): Add basic unit tests for the collector
Following the aws.go example, this adds unit tests for the collector.
The main area where it diverges is that the CollectMetrics doesn't
return an error, so we can't really test for that. Short term I'm okay
with that, long term I believe we'll refactor this to call a `Scrape`
method that returns 0 or 1 and test for that condition being returned.