fix(s3): defensive programming in unitCostForComponent
The previous change to fix the flaky test resulted in breaking other tests.
In `s3.AddMetricGroup` we iterated over all the metrics passed in and at the end of each metric would make a call to `unitCostForComponent`.
Some of the tests were setup in such a way where `unitCostForComponent` was never called, resulting in the PricingMap to have the default float64 value(0).
When moving the call to `unitCostForComponent` to the end of `s3.AddMetricGroup`, this ensured it would get called exactly one time.
This exposed a flaw in `unitCostForComponent` where a usage value of 0 would return NaN instead of 0.
To address this, `unitCostForComponent` will first check if `pricing.Usage` is 0 and then return 0 if so.