fix(s3): Implement billingData cache
Adds a caching layer for `billingData` for S3 to ensure we can export
metrics between calls to gather pricing data.
Renames `S3BillingData` to simply `BillingData` to adhere to Go's
recommendations on naming(IE, not prefixing a type with the package name).
Makes `BillingData` a pointer so that we can do nil checking in `Collect`.
Refactors the chain for calls in `Collect` to be:
- `getBillingData`
- `exportMetrics`

Removes the `skip collection` test from the `Collect` suite as the test
is redundant now that we explicitly test for this scenario.