refactor(providers): Remove CollectMetrics from providers
Remove the CollectMetrics from providers all together and fixes the test cases that failed in the previous commit.
There's still some work to do for the tests so that they're meaningful.
Specifically, I want to add up like metrics to the providers that emits an up metric for each collector.