refactor(providers): Implement operational metrics as consts
Now that the providers accept a channel, we can start to instrument our providers with constant metrics.
The first we're adding is a collectorSuccess gauge that tracks whether each Collector's collect call is successful.
The tests need to be modified slightly to read the channel passed in to each collect method.
Still to do is to validate that the appropriate metrics are returned.