refactor(providers): Update CollectMetrics to pass in a channel
This is the next phase in enabling us to create new metrics everytime we
tick. The most obvious place to start was gcp's compute module as it
contains a bug. If an instance is deleted or no longer exists, the
module still stores the previous values in the
Instance(CPU|Memory)HourlyCost objects. The suggested solution in this
instance is to emit a new const metric each time CollectMetrics is
called and send the Metric to the channel passed in.

This does not remove the existing metrics, as I need to figure out how
to update our testing methodology.