refactor(provider): Expand Provider interface to have Collect methods (#69)
This updates the provider and collector interfaces to properly implement the `prometheus.Collector` interface. 
Updates `exporter.go` to register the provider directly with the registry and ensure it's collect method gets called when Prometheus handles a `/metric` request.
Implements the `Collect` method on each provider in such a way to collect metrics from each Collector in parallel and emit a success metric. 
Refactors `pkg/google/compute/compute.go` to remove metrics stored as global variables in favor of emitting a new metric each time Collect is called. 
As part of refactoring the compute module, update unit tests to test out directly against the metrics returned from the channel.
Introduced a `utils/metrics.go` file to help with testing out metrics generated and used this for `compute_test.go`. 

This PR does _not_ update all collectors to emit new metrics. 
I tried to implement emitting new metrics in `gcs.go` and failed because of duplicate metrics.
The duplicated metrics is a bug that must be addressed in a separate PR.