feat(gke): Implement basic GKE module
Similar to #83, this is a POC to associate a clustername with a compute
instance. This implementation creates a dedicated module for GKE that
will export a single metric: `cloudcost_exporter_gke_node_info` that has
three labels(that can certainly change):
- project
- instance_name
- cluster_name

Refactors the compute module slightly so that I can reuse the
`ListInstances` method in such a way that adds a `ClusterName` attribute
to MachineSpec.