feat(gke): Implement basic GKE module (#84)
Adds a new collector to the `google` family: `gke_compute`. It largely builds on top of the compute module but with one caveat: It is specifically searching for compute instances that belong to a GKE cluster. This took quite a bit of refactoring to mitigate the amount of code duplication.

High level refactors introduced:
- Move calls to `billing.CloudCatalogClient` to a dedicated module
- Move associated `billing.CloudCatalogClient` helper methods from `compute_test.go` to `billing/billing_test_helper.go` 
- Update `compute.go` and `gcs.go` to use the dedicated module
- Move `MachineSpec` and associated methods to `machinespec.go`
- Simplify instantiation of `gcs` in `gcp.go`

There are two new metrics exported:

```
cloudcost_gcp_gke_node_cpu_usd_per_core_hour
cloudcost_gcp_gke_node_memory_usd_per_gib_hour
```

These share almost all of the same labels, with the addition of a `cluster_name` label which is set to the cluster the instance belongs to. 

Co-authored-by: Erik Sommer <erik.sommer@grafana.com>