chore(go): Upgrade to latest version of go (#113)
One thing I noticed in the #109 is that the linting was failing a condition that has since been fixed in the latest version of go. Specifically loopvar scope problems.

This updates:

    go version to 1.22
    golangci-lint to latest version
    Test cases to remove the unnecessary variables

Also turns out that the CI step to get the go version was broken 😱. I'd love to have a better way of extracting this metadata, but for now it's quick and dirty and works.