Get disk pricing working
This implements a `ListDisks` method in `gke` that is then used to list
out disks for the GKE projects by zone.
`ListDisks` will currently return a raw `compute.Disk` struct that is
filtered by
1. Used disks
2. Disks only associated with GKE clusters

The exported price is assuming the disks costs $0.05/GiB/Month and
does not account for the type of disk.