docs: Add necessary files to go opensource (#133)
This change adds the following files necessary for us to go public:
- LICENSE
- SECURITY
- CODE_OF_CONDUCT
- CONTRIBUTING

I've also moved the running section to `contribute/developer-guide.md` to keep the README.md slimmer.

For the LICENSE, I've opted initially for an Apache license. If we find the need to move over to AGPL then it should be fine.