Create dedicated Disk struct
Part one of a set of changes that will introduce a Disk class that will
make it easier to determine if a Disk is a bootdisk or not.

This introduces the `Disk` struct with necessary attributes to meet
existing requirements. It still uses the helper methods defined in
`gke.go`, which will be moved over in the next change.