Feat/pvc attribute boot disks (#151)
Introduces a new label `disk_type` which can be represented as
- `boot_disk`
- `persistent_volume`

The purpose of this is to help further differentiate the type of disks that are being exported. When comparing exported results, I found ~4000 disks that were not represented by `opencost`. After investigating further, these disks were specifically related to boot disks. 

The neat thing about this is we can start associated the costs of boot volumes to the team that is responsible for provisioning nodes within our k8s environments. 

- relates to #5