fix(gcp): Update cost calculation of PVC's to use GiB
Adds a new method to `Disks` to convert the disk size from GB to GiB and
implements unit tests for said method.
Updates documentation and metric name to clarify that the metric already
takes into account the size and cost of the disk. Other metrics export
the raw cost per hour.

- closes #153