feat(gcp+aws): Add last_scrape_time metric
In order to better track freshness of data, this PR adds a few more
operational metrics:
- `cloudcost_exporter_collector_last_scrape_time`
- `cloudcost_exporter_last_scrape_time`

The intent of these is to export in unix time the last time a scrape was
performed. This can be used to alert in prometheus when the
last_scrape_time was say > 60m.

This also implements in AWS the operational metrics that GCP
implemented so that we have feature parity between the two. In the
future it would make sense to generalize this to a common interface so
that new providers do not need to implement the same metrics.