ci(build): Update build to use make
Previously we relied upon `GO_LDFLAGS` to be passed in to get
information such as the branch, revision, and version information.
Another pattern we can follow is to use the existing `make build-binary`
command which as all of the scripting defined for the necessary
variables to build up the `GO_LDFLAGS`. In order to support this, we
need `git` and `make` installed, which thankfully come bundled with the
regular golang image.

Updates `VERSION` to be `git describe --tags --dirty --always` which
will use the the tag information from git directly.