docs: Refactor base README (#161)
With the push to release the project publicly, I wanted to refactor the base README. For internal usages it was fine, but if it's public I think it's lacking clarity on what exactly the project is trying to solve. This is a first draft on improving the messaging about what the project is, isn't, and how to get started.

There's sections that will be sparse until we have a better story around a helm chart or other release mechanisms. IE, how to operate the thing.

- refs #5