feat(eks): Fetch spot pricing for instances
Adds the ability to fetch and parse out spot prices for instances. This
builds upon the [DescribeSpotPriceHistory](https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeSpotPriceHistory.html) api call.
One of the main challeneges with the API response is that it does not
provide attributes such as CPU and Memory counts, which is necessary for
calculating the cost per core/GiB of ram.
In order to gather that metadata, we first build the pricemap for
ondemand instances and generate a struct that is
`map[string]Attributes`. The key is the instanceType and the value is
`Attributes` which contains vCPU + Memory. When building the Spot
pricing map, we perform a lookup to determine how much vCPU + Memory the
instance has.