Add base level log struct
There are two main requirements for improving the logging of
cloudcost-exporter:
1. Structured logging with consistent labels across providers and
   collectors
2. Ability to define a log level and runtime that limits the logs
   emitted

There are no shortage of existing Go packages that could have been used
to meet the first requirement, such as:
- [logrus](https://pkg.go.dev/github.com/sirupsen/logrus)
- [zap](https://github.com/uber-go/zap)
- [zerolog](https://github.com/rs/zerolog)

However each of these comes with tradeoffs and then yet another
dependency to maintain. The go team released
[slog](https://go.dev/blog/slog) officially in go 1.21, which provides a
structured logging implemention as part of the standard library. This
makes it a prime candidate to immediate meet the first requirement.

The second requirement(levels) is a bit more nuanced since the slog is a
thin implementation. The go team provides a reference example to use:
https://pkg.go.dev/log/slog@master#example-Handler-LevelHandler

This PR aims to meet both requirements by
1. Adding `slog` as a common logging library to use
2. Creating a levels implementation that allows us to toggle the emitted
   level

There are two three methods added as well:
- `GetLogLevel` to convert a string to the log level
- `WriterForOutput` to convert a string to an `io.Writer`
- `HandlerForOutput` that takes a string and returns a slog handler for
  json or text