feat(ec2): Add skeleton of ec2 module (#202)
This starts the process of refactoring `eks` and creating the skeleton for an ec2 module. Since `eks` and `ec2` are going to share a significan amount of code, I've introduced a `compute` module within `aws`. At a bare minimum the `pricing_map.go` will be shared, so that lives within the `compute` module.

Adds a base skeleton implementation for `ec2` that implements the collector interface and has 100% test coverage to start off.

Updates `aws.go` to handle an `ec2` flag, but does not document it's usage just yet.

First step of #200