feat(ec2): Implement generating a pricing map (#211)
This is the next step towards implementing #200.
Updates logic in `pricing_map.go` to add an attribute for the total price of the instance.
Largely copies the implementation in `eks.go` to generate the pricing map by listing out OnDemand and Spot prices.
Refactors shared methods from `eks.go` to list on demand prices, spot prices, and instances to `pricing_map.go` and `compute.go` respectively. Create unit tests to ensure `ec2.Collect` has 100% coverage.

One thing that I'm pretty conscious of and don't have a great way of handling is how similar the `Collect` methods are up to this point. Right now I'm okay with duplicated code between the two as the remaining implementation will differ. I think long term there's an opportunity to refactor how we handle the pricing map that is shared between modules, but for now that's out of scope.