feat(ec2): Export metrics for ec2 instances
This PR ties together the #211 and lists out instances for a specific
AWS profile. The code checks for instances that are not associated with
a cluster and then emits a metric for the total hourly cost.

Adds documentation for how the collector is created and provides some
basic PromQL queries that _could_ be used.