feat(eks): Export total hourly cost for ec2 + eks instances
This removes the if conditione that filtered out ec2 instances that did
not have a cluster name.
The rationale is that after implementing #214, it became apparent that
there is no inherent value in having the two implementations split out.
Prometheus will understand that an empty label is valid and be treated
as if the label does not exist. This has been validated by emitting
a set of production data and having it scraped into a Prometheus
instance and submitted queries against it.

Adds a metric(`cloudcost_aws_eks_total_hourly_cost`) to emit for total hourly cost for each instance.