feat(eks): Export total hourly cost for ec2 + eks instances (#215)
This removes the if conditione that filtered out ec2 instances that did not have a cluster name.
The rationale is that after implementing #214, it became apparent that there is no inherent value in having the two implementations split out. Prometheus will understand that an empty label is valid and be treated as if the label does not exist. This has been validated by emitting a set of production data and having it scraped into a Prometheus instance and submitted queries against it.

Adds a metric(`cloudcost_aws_eks_total_hourly_cost`) to emit for total hourly cost for each instance.

The plan for the PR is to will delete the implemented ec2 module and refactor the existing implementation to be called simply ec2. I think this is closer to reality and will allow us to be able to track the costs of ec2 instances associated with eks clusters and those that don't.