docs(developer-guide): Explicitly list external dependencies
There are two external tools used by the project that are not defined by
`go.mod`:
1. mockery
2. mockgen

This adds documentation on where to look for installation instructures
and specifies to use the latest version. The primary reason why I
suggest latest instead of the current version used is because PR's
generated by the latest tools can serve as a reminder for others to
upgrade their tooling. I don't know of a way we can automate the upgrade
of those tools.