chore(GCP): Delete compute module (#317)
When initially building out the compute module, we had an idea that we wanted to keep the GKE/Compute module completely seperate. The primary motivation was because the labels would be different and thus we would want distinct metric names. Since then, Grafana has not moved the compute module into production and we have since implemented a similar feature for AWS. The main decision we made was to have the `cluster_name` label empty, which Prometheus will then discard.

This change removes the compute module, but does _not_ implement parity. We can follow up with that once there is a demand for us to develop a module specifically for GCE instances.