fix(gke): Update collect method to gracefully fail
There is a condition where a failure to find a machine type in the price
map was causing the `Collect` method to return instead of continuing.
Instead, this changes the logic to log out the message within `Collect`
and continue operating.
Adds a unit test to reproduce the error and then validates it is
resolved.

A future improvement here would be to
1. Export a metric
2. Set a sane default value(if provided?)

- closes #335