feat(gke): Handle hyperdisk balanced volumes
First attempt at handling hyperdisk balanced volumes.
The primary goal was to update the pricing map logic in such a way that would parse of hyperdisks  balanced and include only the cost of capacity.
Hyperdisk Balanced pricing is done in such a way that if you don't configure IOPS and Throughput, you use a free tier.

In the future IOPS, Throughput, and High Availability costs need to be saved and calculated as well.