feat(ec2): Add architecture label
During capacity planning it's often useful to be able to answer a
question such as the count of ec2 instances by machine type that are of
a particular architecture. A simple example would be

```promql
sum by (machine_type, cluster_name) (
  cloudcost_aws_ec2_instance_cpu_usd_per_core_hour{architecture="arm64"}
)
```

Updates the labels for the ec2 cost metrics to include an `architecture`
label.