Dashboard Builder Experiment
One of the goals is to be able to consume the dashboards as if it they
were a library so that they can be built in other contexts, such as
internal repositories that manage dashboards.

My proposal would be for there to be a way of getting all of the
dashboards a project can export as a map where the key is the dashboard
name, and the value is the string representation of the dashboard. This
way a downstream application can do something like

```golang
package main
import (
  "github.com/grafana/cloudcost-exporter/dashboards"
)

func main() {
  for name, content := range dashboards {
     // write out content to ${name}.json
  }
}
```