Create script to emulate calculate costs of persistent volumes
Implements the foundational parts of calculating the costs of persistent
volumes for Azure disks. There's a README.md file that is a rough
scratch of my development process that will be refined prior to merging
in.

This PR implements three core ideas
- Download pricing fact sheets from Azure via it's retail price SDK and
  converts data into a pricing map
- Lists out disks that are running in a specific subscription
- A script that combines these two elements to output the cost of a disk

There's one major flaw in this, in that getting the price for a specific
disk does not work properly. The data returned from Azure is disjointed
and it's not a simple 1:1 mapping of a disk sku to a pricing sku. This
needs to be solved, and to do that I've put together a (private)
spreadsheet that uses data from both sources.

- relates to #236