* Migrate NATGateway from Cost Explorer to Pricing API
This commits refactors the NAT Gateway collector to use the Pricing API
instead of the Cost Explorer API. This is consistent with other services
such as EC2.

Reason: Cost Explorer API calls charge us $0.01 per call, whereas Price List API
calls are free of charge.

Additional: The pricing store for NAT Gateways is purposely built to be as
product-neutral as possible so that it can be refactored into its own
package and help us ship new similar services faster.